renamed source files to lowercase
