+++ src/tools/qgraphicscharaitem.cpp
