+++ src/tools/qgraphicspaletescene.h
